Dependencies: Add libiconv and libintl
libiconv should be at /usr/lib/ on most systems but could be a wrong version and often gets messed up and libintl is not on a clean install so need to bundle them with VDrift. Haven't been able to build these myself, so they are left as dylibs. Had to change the install names (otool -id etc) to make sure they work on other people's systems, and code sign them (codesign -fs - etc). Tested on 10.7 upwards.